refractor: some UX improvements

+++ libs/word_struct.c
@@ -1,7 +1,7 @@
 #include "word_struct.h"
 
-unsigned int string_hash(const char *s, unsigned long SIZE) {
-    unsigned int hash = 5381;
+unsigned long string_hash(const char *s, unsigned long SIZE) {
+    unsigned long hash = 5381;
     while (*s)
         hash = 33 * hash ^ (unsigned char) *s++;
     return hash % SIZE;
